requirements.txt
cryptography==35.0.0
Pillow==8.4.0